% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vtk.R
\name{read_vtk_points}
\alias{read_vtk_points}
\title{Read VTK points}
\usage{
read_vtk_points(x, ..., quiet = FALSE)
}
\arguments{
\item{x}{VTK file name}

\item{...}{ignored}

\item{quiet}{if `FALSE` (the default), refrain from reporting extra information}
}
\value{
matrix of x, y, z coordinates
}
\description{
Read a VTK (Visualization Toolkit ) STRUCTURED GRID of POINTS.
}
\details{
Only BINARY supported for now.

Note that we currently do not recreate the matrix of data, just return the points. The
header is printed if `verbose = TRUE` so the dimensions reported there may be used.

Specification of the format found at \url{https://vtk.org/wp-content/uploads/2015/04/file-formats.pdf}
on 2019-10-09.
}
\examples{
f <- system.file("extdata/STRUCTURED_GRID/volcano_points.vtk", package = "vtkr", mustWork = TRUE)
xyz <- read_vtk_points(f)
}
