% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtkr-package.R
\docType{package}
\name{vtkr-package}
\alias{vtkr}
\alias{vtkr-package}
\title{vtkr: Read a Limited Subset of Visualization Toolkit Formats}
\description{
Read some VTK formats. VTK is the Visualization Toolkit <https://en.wikipedia.org/wiki/VTK>.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
